Correção e Formatação

+++ categoria_1/1038.py
@@ -1,5 +1,4 @@
-
-X, Y = map(int, input().split())
+X, Y = map(float, input().split())
 
 if X == 1:
     res = 4 * Y
